Successfully run pytest on 'test_check_stop_loss'.

+++ src/strat_backtest/stop_method/nearest_loss.py
@@ -53,4 +53,4 @@ class NearestLoss(StopLoss):
         # for short position
         stop_price = max(stop_list) if entry_action == "buy" else min(stop_list)
 
-        return Decimal(round(stop_price, 2))
+        return round(stop_price, 2)
